add a few scenarios to benchmark
