+++ Benchmark/Benchmark.V6/BasicUsages.cs
@@ -10,7 +10,7 @@ using Reactive.Bindings.Extensions;
 
 namespace ReactivePropertyBenchmark
 {
-    public class BasicUsages
+    public partial class BasicUsages
     {
         [Benchmark]
         public ReactiveProperty<string> CreateReactivePropertyInstance() => new ReactiveProperty<string>();
@@ -52,6 +52,13 @@ namespace ReactivePropertyBenchmark
         {
             var p = new Person();
             return p.ToReactivePropertyAsSynchronized(x => x.Name);
+        }
+
+        [Benchmark]
+        public ReactivePropertySlim<string> ToReactivePropertyAsSynchronizedSlim()
+        {
+            var p = new Person();
+            return p.ToReactivePropertySlimAsSynchronized(x => x.Name);
         }
     }
